% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_validate.r
\name{biclust_smooth}
\alias{biclust_smooth}
\title{Bicluster Smooth}
\usage{
biclust_smooth(X, clusters_row, clusters_col)
}
\arguments{
\item{X}{Data matrix}

\item{clusters_row}{Row cluster object}

\item{clusters_col}{Column cluster object}
}
\description{
\code{biclust_smooth} computes the bicluster estimates given row and column partitions
}
\seealso{
\code{\link[cvxbiclustr]{get_subgroup_means_full}}
}
