% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra.r
\name{COBRA_missing}
\alias{COBRA_missing}
\title{Algorithm 2 of Convex Biclustering by Chi et al.}
\usage{
COBRA_missing(
  X,
  W_row,
  W_col,
  gamma = 10,
  tmax_hierarchy = c(50, 10, 1000),
  tol = 1e-04,
  progress = FALSE
)
}
\arguments{
\item{X}{data matrix}

\item{W_row}{\code{sparseweights} object output from \code{knn_graph} for row fusion}

\item{W_col}{\code{sparseweights} object output from \code{knn_graph} for col fusion}

\item{gamma}{regularization parameter}

\item{tmax_hierarchy}{vector of 3 values
\enumerate{
\item number of COBRA fits
\item number of number of Convex Clustering fits per COBRA fit
\item number of iterations per Convex Clustering fit
}}

\item{tol}{of convergence}

\item{progress}{show progress}
}
\value{
list with info
4. \code{valid_rss} unweighted RSS of non-missing values
6. \code{filled_vals} imputed values from each COBRA fit
7. \code{invalid_indices} indices of missing values
8. \code{time} to run
}
\description{
Algorithm 2 of Convex Biclustering by Chi et al.
}
