% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_cv.R
\name{cv.BCBC_holdout}
\alias{cv.BCBC_holdout}
\title{Title}
\usage{
cv.BCBC_holdout(
  X,
  holdout_size = 0.1,
  lambdas = c(1),
  gammas = c(1),
  k_rows = c(2),
  k_cols = c(2),
  phis = c(0.5),
  recalculate_weights = c(FALSE),
  tols = c(0.001),
  tmax_outers = c(50),
  tmax_inners = c(50),
  retrain = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
