% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_cv.R
\name{cv.BCBC_holdout_stability}
\alias{cv.BCBC_holdout_stability}
\title{An alternative method to select a value for lambda by checking if the clusters
from BCBC match the clusters from COBRA run on the weighted matrix.}
\usage{
cv.BCBC_holdout_stability(
  X,
  gamma = 1,
  lambdas = c(0),
  k_samples = 2,
  k_features = 2,
  phi = 0.5,
  recalculate_weights = TRUE,
  tol = 1e-04,
  tmax_outer = 50,
  tmax_inner = 50,
  reweight = TRUE,
  percent_noise = 0.25,
  method = "adjusted.rand",
  progress = TRUE
)
}
\arguments{
\item{recalculate_weights}{}

\item{reweight}{reweight the columns before passing into COBRA}
}
\description{
An alternative method to select a value for lambda by checking if the clusters
from BCBC match the clusters from COBRA run on the weighted matrix.
}
