% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{get_cv_metrics}
\alias{get_cv_metrics}
\title{Get metrics to measure a single BCBC fit that are useful for cross validation}
\usage{
get_cv_metrics(
  X,
  bcbc_run,
  percent_noise = 0.25,
  num_row_clusters = NA,
  num_col_clusters = NA
)
}
\arguments{
\item{percent_noise}{Percent of standard deviation of pairwise distances to use for thresholding}

\item{num_row_clusters}{If set, tries to calculate a threshold to result in this many row clusters}

\item{num_col_clusters}{If set, tries to calculate a threshold to result in this many column clusters}
}
\description{
Get metrics to measure a single BCBC fit that are useful for cross validation
}
