% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_row_clusters}
\alias{get_row_clusters}
\title{Get row clusters memberships with thresholding radius.}
\usage{
get_row_clusters(mat, threshold)
}
\arguments{
\item{mat}{to find row memberships for}

\item{threshold}{radius to join rows into communities}
}
\value{
vector of row memberships
}
\description{
Get row clusters memberships with thresholding radius.
}
\examples{
checker <- gen_checkerboard(100, 150, 100, 5, shuffle = TRUE)
get_row_clusters(checker$X, 10*sd(checker$X)^2)
}
\seealso{
\code{\link[dbscan:frNN]{dbscan::frNN()}}
}
