% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_validate.r
\name{get_subgroup_means_full}
\alias{get_subgroup_means_full}
\title{Get Subgroup Means}
\usage{
get_subgroup_means_full(X, clusters_row, clusters_col)
}
\arguments{
\item{X}{Data matrix}

\item{clusters_row}{Row cluster object}

\item{clusters_col}{Column cluster object}
}
\description{
\code{get_subgroup_means_full} computes the subgroup means.
}
\seealso{
\code{\link[cvxbiclustr]{get_subgroup_means}}
}
