% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_util.r
\name{gkn_weights}
\alias{gkn_weights}
\title{Gaussian Kernel + k-Nearest Neighbor Weights}
\usage{
gkn_weights(X, phi = 0.5, k_row = 5, k_col = 5, return_connectivity = TRUE)
}
\arguments{
\item{X}{The data matrix to be clustered. The rows are the features, and the columns are the samples.}

\item{phi}{The nonnegative parameter that controls the scale of kernel weights}

\item{k_row}{The number of row nearest neighbors}

\item{k_col}{The number of column nearest neighbors}

\item{return_connectivity}{calculate and connectivity information for X}
}
\description{
\code{gkn_weights} combines Gaussian kernel weights with k-nearest neighbor weights
}
