% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{knn_graph_approx}
\alias{knn_graph_approx}
\title{Calculate approximate nearest neighbors for use with \code{BCBC}, \code{cobra} and subroutines.}
\usage{
knn_graph_approx(
  X,
  k,
  sum_to_one = TRUE,
  rescale_by_dim = ncol(X),
  uniform = FALSE,
  phi = 1/ncol(X),
  hnsw_args = list()
)
}
\arguments{
\item{X}{data matrix}

\item{k}{number of nearest neighbors}

\item{sum_to_one}{make weights sum to one before any dimension rescaling}

\item{rescale_by_dim}{rescale weights by dividing by sqrt(this number) (if finite)}

\item{uniform}{treat all edges as equal magnitude}

\item{phi}{bandwidth for Gaussian kernel}

\item{hnsw_args}{passed to \code{RcppHNSW::hnsw_knn}}
}
\value{
\code{sparseweights} object (list) with
\enumerate{
\item \code{keys} matrix of edges
\item \code{values} weights for edges in \code{keys}
\item \code{fusion} sum of fusion term when using the weights and distances
}
}
\description{
Calculate approximate nearest neighbors for use with \code{BCBC}, \code{cobra} and subroutines.
}
