% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_fit_to_df}
\alias{matrix_fit_to_df}
\title{Title}
\usage{
matrix_fit_to_df(
  matrix,
  labels = NULL,
  labels2 = NULL,
  weights = c(),
  lambda = 0,
  filter_weight = 0,
  cluster_with_weights = TRUE
)
}
\arguments{
\item{cluster_with_weights}{}
}
\description{
Title
}
