% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbc.R
\name{projection_onto_simplex}
\alias{projection_onto_simplex}
\title{Orthogonal projection of a vector onto the simplex}
\usage{
projection_onto_simplex(unproj, bound = 1)
}
\arguments{
\item{unproj}{input vector}

\item{bound}{sum of projected vector}
}
\value{
orthogonal projection of \code{unproj}
}
\description{
From https://github.com/kharchenkolab/vrnmf/tree/main
}
\examples{
projection_onto_simplex(c(0.3, 0.3, 0.2))

}
