% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_compare}
\alias{safe_compare}
\title{Method to easily use \code{igraph::compare}}
\usage{
safe_compare(estimated_clusters, true_clusters, method)
}
\arguments{
\item{estimated_clusters}{community memberships}

\item{true_clusters}{community memberships}

\item{method}{passed to \code{igraph::compare}}
}
\value{
metric in community memberships
}
\description{
Method to easily use \code{igraph::compare}
}
\seealso{
\code{\link[igraph:compare]{igraph::compare()}}
}
