% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{thresholded_solution}
\alias{thresholded_solution}
\title{Title}
\usage{
thresholded_solution(bcbc_result, percent_of_noise, cluster_w_weights = TRUE)
}
\arguments{
\item{cluster_w_weights}{}
}
\description{
Title
}
