% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{unweighted_bicluster_assignments}
\alias{unweighted_bicluster_assignments}
\title{Title}
\usage{
unweighted_bicluster_assignments(
  U,
  num_row_clusters = NA,
  num_col_clusters = NA,
  percent_noise = 0.25
)
}
\arguments{
\item{percent_noise}{}
}
\description{
Title
}
