% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbc.R
\name{BCBC}
\alias{BCBC}
\title{Title}
\usage{
BCBC(
  X,
  lambda,
  k_features = 4,
  k_samples = 4,
  gamma = 10,
  phi = 0.05,
  tmax = NA,
  tmax_cobra = 100,
  tmax_outer = 100,
  tol = 1e-05,
  recalculate_weights = TRUE,
  greedy_terminate = !recalculate_weights,
  approx = 0,
  progress = TRUE
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
