% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{BCBC_missing}
\alias{BCBC_missing}
\title{Title}
\usage{
BCBC_missing(
  X,
  lambda,
  k_row = 4,
  k_col = 4,
  gamma = 10,
  phi = 0.05,
  tmax_inner = 50,
  tmax_outer = 100,
  tol = 1e-04,
  recalculate_weights = FALSE,
  approx = 0,
  progress = TRUE,
  fusion_weights = NA
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
