% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{BCBC_missing}
\alias{BCBC_missing}
\title{Title}
\usage{
BCBC_missing(
  X,
  lambda,
  k_samples = 4,
  k_features = 4,
  gamma = 10,
  phi = 0.05,
  tmax_hierarchy = c(10, 25, 50),
  tol = 1e-04,
  recalculate_weights = FALSE,
  approx_neighbors = FALSE,
  hnsw_args = list(),
  progress = FALSE,
  fusion_weights = NA,
  return_fit = FALSE
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
