% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bcbc_result_to_df}
\alias{bcbc_result_to_df}
\title{Title}
\usage{
bcbc_result_to_df(
  bcbc_result,
  cluster_with_weights = TRUE,
  filter_weight = -Inf,
  labels = NULL,
  labels2 = NULL
)
}
\arguments{
\item{labels2}{}
}
\description{
Title
}
