% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{bicluster_assignments}
\alias{bicluster_assignments}
\title{Title}
\usage{
bicluster_assignments(
  U,
  weights = NA,
  lambda = NA,
  num_row_clusters = NA,
  num_col_clusters = NA,
  percent_noise = 0.25
)
}
\arguments{
\item{percent_noise}{}
}
\description{
Title
}
