% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{build_weights_for_edges}
\alias{build_weights_for_edges}
\title{Title}
\usage{
build_weights_for_edges(
  edges,
  distances,
  uniform = FALSE,
  rescale = 1,
  phi = 1
)
}
\arguments{
\item{phi}{}
}
\description{
Title
}
