% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{build_weights_for_edges}
\alias{build_weights_for_edges}
\title{Build list for usage with \code{cobra}.}
\usage{
build_weights_for_edges(
  edges,
  distances,
  sum_to_one = TRUE,
  rescale_by_dim = NA,
  uniform = FALSE,
  phi = 1
)
}
\arguments{
\item{edges}{id attribute matrix output from \code{dbscan::kNN} or (idx) \code{RcppHNSW::hnsw_knn}}

\item{distances}{dist attribute output from \code{dbscan::kNN} or \code{RcppHNSW::hnsw_knn}}

\item{sum_to_one}{make weights sum to one before any dimension rescaling}

\item{rescale_by_dim}{rescale weights by dividing by sqrt(this number) (if finite)}

\item{uniform}{treat all edges as equal magnitude}

\item{phi}{bandwidth for Gaussian kernel}
}
\value{
\code{sparseweights} object (list) with
\enumerate{
\item \code{keys} matrix of edges
\item \code{values} weights for edges in \code{keys}
\item \code{fusion} sum of fusion term when using the weights and distances
}
}
\description{
Build list for usage with \code{cobra}.
}
