% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbc.R
\name{calc_fusion_term}
\alias{calc_fusion_term}
\title{Calculate the fusion term in the BCBC objective function}
\usage{
calc_fusion_term(edges, U, rows = TRUE)
}
\arguments{
\item{edges}{\code{sparseweights} object}

\item{U}{input to fusion function}

\item{rows}{logical value indicating to calculate over rows}
}
\value{
fusion value
}
\description{
Calculate the fusion term in the BCBC objective function
}
\examples{
checker <- gen_checkerboard(n = 100,
                            p = 100,
                            num_row_clusters = 5,
                            num_col_clusters = 5,
                            p_extra = 25,
                            shuffle = FALSE)
fusion_wts <- fast_gkn_weights(checker$X, k_row = 4, k_col = 4)
calc_fusion_term(fusion_wts$W_row, checker$X, rows=FALSE)
calc_fusion_term(fusion_wts$W_col, checker$X, rows=TRUE)
}
\seealso{
\code{\link[=fast_gkn_edges]{fast_gkn_edges()}} for Gaussian k-nearest neighbor affinities

\code{\link[=knn_graph]{knn_graph()}} for Gaussian k-nearest neighbor affinities
}
