% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra.r
\name{cobra}
\alias{cobra}
\title{Dykstra-like proximal algorithm from Convex Biclustering by Chi et al.}
\usage{
cobra(
  X,
  gamma,
  W_row,
  W_col,
  max_iter = 100,
  max_iter_inner = 100,
  tol = 1e-04
)
}
\arguments{
\item{X}{data matrix}

\item{gamma}{regularization parameter}

\item{W_row}{\code{sparseweights} object output from \code{knn_graph} for row fusion}

\item{W_col}{\code{sparseweights} object output from \code{knn_graph} for col fusion}

\item{max_iter}{number of convex cluster optimizations to solve}

\item{max_iter_inner}{number of inner iterations for \code{CCMMR}}

\item{tol}{of convergence}
}
\value{
fitted bicluster matrix
}
\description{
Dykstra-like proximal algorithm from Convex Biclustering by Chi et al.
}
\examples{
checker <- gen_checkerboard(100, 150, 5, 5, p_extra = 0, shuffle = FALSE)
W_row <- knn_graph(checker$X, 5)
W_col <- knn_graph(t(checker$X), 5)
cobra_fit <- cobra(checker$X, gamma = 10000, W_row = W_row, W_col = W_col)

plot_fit(matrix_fit_to_df(cobra_fit))
}
\seealso{
\code{\link[=fast_gkn_weights]{fast_gkn_weights()}}

\code{\link[=knn_graph]{knn_graph()}}

\code{\link[=knn_graph_approx]{knn_graph_approx()}}
}
