% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.BCBC}
\alias{cv.BCBC}
\title{Title}
\usage{
cv.BCBC(
  X,
  lambdas = c(1),
  k_rows = c(2),
  k_cols = c(2),
  gammas = c(1),
  tmaxs = c(NA),
  tmax_outers = c(100),
  tmax_cobras = c(100),
  phis = c(0.5),
  recalculate_weights = c(TRUE),
  tols = c(1e-06),
  weighted_clusters = TRUE,
  percent_noise = c(0.1),
  progress = FALSE
)
}
\arguments{
\item{progress}{}
}
\description{
Title
}
