% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.BCBC}
\alias{cv.BCBC}
\title{Title}
\usage{
cv.BCBC(
  X,
  lambdas = c(1),
  k_samples = c(2),
  k_features = c(2),
  gammas = c(1),
  tmaxs = c(NA),
  tmax_outers = c(100),
  tmax_cobras = c(100),
  phis = c(0.5),
  recalculate_weights = c(TRUE),
  tols = c(1e-06),
  percent_noise = c(0.25),
  num_row_clusters = NA,
  num_col_clusters = NA,
  ...
)
}
\arguments{
\item{...}{passed to BCBC}
}
\description{
Title
}
