% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holdout_cv.R
\name{cv.BCBC_holdout}
\alias{cv.BCBC_holdout}
\title{Title}
\usage{
cv.BCBC_holdout(
  X,
  holdout_size = 0.1,
  lambdas = c(1),
  gammas = c(1),
  k_samples = c(2),
  k_features = c(2),
  phis = c(0.5),
  recalculate_weights = c(FALSE),
  approx_neighbors = c(FALSE),
  hnsw_args = list(),
  tols = c(0.001),
  tmax_hierarchy = c(10, 25, 50),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
