% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{fast_gkn_weights}
\alias{fast_gkn_weights}
\title{Build row and column info for usage with \code{BCBC}, \code{cobra} and subroutines.}
\usage{
fast_gkn_weights(
  X,
  k_row,
  k_col,
  phi,
  approximate = FALSE,
  rescale = TRUE,
  uniform = FALSE,
  hnsw_args = list(ef = 50)
)
}
\arguments{
\item{X}{data matrix}

\item{k_row}{number of nearest neighbors for rows}

\item{k_col}{number of nearest neighbors for columns}

\item{phi}{bandwidth for Gaussian kernel}

\item{approximate}{use RcppHNSW for calculation}

\item{rescale}{weights and bandwidth by this factor}

\item{uniform}{treat all edges as equal magnitude}

\item{hnsw_args}{passed to \code{RcppHNSW::hnsw_knn}}
}
\value{
list of row and column info similar to \code{knn_graph} return info
}
\description{
Build row and column info for usage with \code{BCBC}, \code{cobra} and subroutines.
}
\examples{
checker <- gen_checkerboard(n = 100,
                            p = 100,
                            num_row_clusters = 5,
                            num_col_clusters = 5,
                            p_extra = 25,
                            shuffle = FALSE)
fusion_wts <- fast_gkn_weights(t(checker$X), k_row = 4, k_col = 4, phi = 1)
calc_fusion_term(fusion_wts$unique_row_edges, fusion_wts$w_row, checker$X, rows=FALSE)
calc_fusion_term(fusion_wts$unique_col_edges, fusion_wts$w_col, checker$X, rows=TRUE)
}
\seealso{
\code{\link[=knn_graph]{knn_graph()}}

\code{\link[=knn_graph_approx]{knn_graph_approx()}}
}
