% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{fast_gkn_weights}
\alias{fast_gkn_weights}
\title{Title}
\usage{
fast_gkn_weights(X, k_row, k_col, phi, approx = 0, return_edges = TRUE)
}
\arguments{
\item{return_edges}{}
}
\description{
Title
}
