% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{full_connectivity}
\alias{full_connectivity}
\title{Build completely connected affinity graph}
\usage{
full_connectivity(
  X,
  sum_to_one = TRUE,
  rescale_by_dim = TRUE,
  uniform = FALSE,
  phi = 1/ncol(X)
)
}
\arguments{
\item{X}{data matrix}

\item{sum_to_one}{sum the weights to one}

\item{rescale_by_dim}{divide the weights by sqrt(ncol(X))}

\item{uniform}{keep all weights uniform}

\item{phi}{exponential kernel scaling}
}
\value{
sparseweights object
}
\description{
Build completely connected affinity graph
}
