% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkerboard.R
\name{gen_checkerboard}
\alias{gen_checkerboard}
\title{Generate simulated data resembling a checkerboard}
\usage{
gen_checkerboard(
  n,
  p,
  num_row_clusters,
  num_col_clusters,
  p_extra = 0,
  noise = 1,
  prob_empty = 0.1,
  shuffle = TRUE,
  scale = TRUE
)
}
\arguments{
\item{n}{number of rows}

\item{p}{number of columns}

\item{p_extra}{number of extra noisy columns}

\item{noise}{standard deviation of normal noise added to checkerboard}

\item{prob_empty}{between 0 and 1, probability a bicluster is empty}

\item{shuffle}{randomly reorder rows and columns}

\item{scale}{the resulting matrix}
}
\description{
Generate simulated data resembling a checkerboard
}
