% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_threshold_for_k_components}
\alias{get_threshold_for_k_components}
\title{Get radius to create k connected components in a nearest neighbor graph}
\usage{
get_threshold_for_k_components(
  dist_mat,
  k,
  max_threshold = max(dist_mat),
  maxiter = 50
)
}
\arguments{
\item{maxiter}{}
}
\description{
Get radius to create k connected components in a nearest neighbor graph
}
