% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{knn_graph_approx}
\alias{knn_graph_approx}
\title{Title}
\usage{
knn_graph_approx(
  X,
  k,
  rescale = FALSE,
  uniform = FALSE,
  phi = 1,
  byrow = FALSE,
  hnsw_args = list()
)
}
\arguments{
\item{hnsw_args}{}
}
\description{
Title
}
