% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_fit}
\alias{plot_fit}
\title{Title}
\usage{
plot_fit(
  plot_df,
  title = "",
  xlabel = "",
  ylabel = "",
  fill_attr = "value",
  alpha_weight = FALSE,
  bin_scale = FALSE
)
}
\arguments{
\item{bin_scale}{}
}
\description{
Title
}
