% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbc.R
\name{projection_onto_simplex}
\alias{projection_onto_simplex}
\title{Orthogonal projection of a vector onto the simplex}
\usage{
projection_onto_simplex(unproj, bound = 1)
}
\arguments{
\item{bound}{}
}
\description{
From https://github.com/kharchenkolab/vrnmf/tree/main
}
