% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcbc.R
\name{w_coordinate_descent}
\alias{w_coordinate_descent}
\title{Title}
\usage{
w_coordinate_descent(col_sum_sq, lambda)
}
\arguments{
\item{col_sum_sq}{Sum of squared errors for columns}

\item{lambda}{}
}
\description{
Title
}
